Duck RNN v2: remove conv stack, add benchmark (#116)
Summary:\n- Remove conv layers from DuckNeuralNetRecurrentBrainV2; feed
raw 21x21x10 histograms + scalars directly into recurrent head.\n-
Switch duck material histograms to float and template SensoryUtils
histogram helpers accordingly.\n- Add a disabled manual benchmark
harness for brain/sensory timing.\n- Fix a couple release-only -Werror
warnings.\n\nNotes:\n- This changes genome layout/weight count for
DuckNeuralNetRecurrentBrainV2; existing genomes/training sessions using
the old conv layout will be incompatible and need
restart/regeneration.\n- Local release benchmark on this branch measured
~3.6us/op brain.think vs ~122us/op for the conv-based version
(order-of-magnitude speedup).

+++ apps/src/core/organisms/DuckSensoryData.h
@@ -20,7 +20,7 @@ struct DuckSensoryData {
     static constexpr int SPECIAL_SENSE_COUNT = 32;
 
     // Material histogram grid: [y][x][material] = fill contribution.
-    std::array<std::array<std::array<double, NUM_MATERIALS>, GRID_SIZE>, GRID_SIZE>
+    std::array<std::array<std::array<float, NUM_MATERIALS>, GRID_SIZE>, GRID_SIZE>
         material_histograms = {};
 
     // Mapping from neural grid to world coordinates.
